Begin TypeScript migration
