+++ package.json
@@ -11,7 +11,8 @@
   },
   "devDependencies": {
     "eslint": "^8.7.0",
-    "nodemon": "^2.0.15"
+    "nodemon": "^2.0.15",
+    "typescript": "^4.6.2"
   },
   "name": "awesomescibo",
   "version": "3.2.1",
