+++ src/helpers/env.ts
@@ -1,0 +1,6 @@
+import 'dotenv/config';
+
+export const clientId : string = process.env.CLIENT_ID!;
+export const testingGuild : string = process.env.TESTING_GUILD!;
+export const token : string = process.env.TOKEN!;
+export const mongoUri : string = process.env.MONGO_URI!;
